requirements.txt
# Core
streamlit>=1.25.0
python-dotenv>=1.0.0
pandas>=2.0.0

# LangChain & OpenAI
langchain>=0.2.9
langchain-openai>=0.0.21
langchain-chroma>=0.0.11
langchain-community>=0.0.8

# Vector DB
chromadb>=0.3.30

# Excel and unstructured document support
openpyxl>=3.1.0
unstructured[local-inference]>=0.7.7

# Optional helpers
tqdm>=4.66.1

# 

